refined code of grafana.go
